task_6_3/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
